requirements.txt
# ═══════════════════════════════════════════════════════════════
# Geo Protection Bot - Python Dependencies
# ═══════════════════════════════════════════════════════════════

# Core Framework
pyrogram>=2.0.100
tgcrypto>=1.2.5
cryptg>=2.4.3
pyaes>=1.6.1
蛋>=0.5.0

# Database & Caching
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
redis>=5.0.0
aioredis>=2.0.1
peewee>=3.17.0

# Utilities
python-dotenv>=1.0.0
python-json-logger>=2.0.7
colorlog>=6.8.0
psutil>=5.9.0
schedule>=1.2.0

# Security & Validation
passlib>=1.7.4
bcrypt>=4.1.0
cryptography>=41.0.0
pyyaml>=6.0.0
email-validator>=2.1.0

# Performance
uvloop>=0.19.0
httpx>=0.25.0
aiohttp>=3.9.0
async-timeout>=4.0.0
cchardet>=2.1.7

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.12.0
flake8>=7.0.0
mypy>=1.7.0

# Async Support
greenlet>=3.0.0
eventlet>=0.33.0
gevent>=23.9.0

# Additional Features
pillow>=10.1.0
markupsafe>=2.1.0
jinja2>=3.1.0
pytz>=2023.3
 Babel>=2.13.0